Moved AggCcsRecordAggregationHelper to own file
